requirements.txt
# Core LangChain
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
langchain-openai>=0.0.2
langchain-qdrant>=0.2.0
langchain-huggingface>=0.0.3
langgraph>=0.0.10

# Vector Store
qdrant-client>=1.10.0

# PDF Processing
pymupdf>=1.24.0
pypdf>=4.0.0

# Ollama
langchain-ollama>=0.1.0

# Embeddings (HuggingFace)
sentence-transformers>=2.6.0
transformers>=4.38.0
huggingface-hub>=0.34.0,<1.0
torch>=2.2.0
accelerate>=0.26.0
sentencepiece>=0.2.0
safetensors>=0.4.2

# LLM Backend
vllm>=0.6.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Testing & HTTP
requests>=2.31.0

python-dotenv>=1.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.8.0